+++ event/preview.php
@@ -11,6 +11,7 @@
 namespace vse\topicimagepreview\event;
 
 use phpbb\config\config;
+use phpbb\template\template;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
 /**
@@ -22,7 +23,10 @@ class preview implements EventSubscriberInterface
 	protected $config;
 
 	/** @var helper */
-	private $factory;
+	protected $helper;
+
+	/** @var template */
+	protected $template;
 
 	/**
 	 * {@inheritdoc}
@@ -30,6 +34,7 @@ class preview implements EventSubscriberInterface
 	public static function getSubscribedEvents()
 	{
 		return [
+			'core.page_footer'						=> 'init_tpl_vars',
 			// Viewforum events
 			'core.viewforum_modify_topics_data'		=> 'viewforum_row',
 			'core.viewforum_modify_topicrow'		=> 'viewforum_tpl',
@@ -48,10 +53,22 @@ class preview implements EventSubscriberInterface
 	 * @param config $config
 	 * @param helper $helper
 	 */
-	public function __construct(config $config, helper $helper)
+	public function __construct(config $config, helper $helper, template $template)
 	{
 		$this->config = $config;
-		$this->factory = $helper;
+		$this->helper = $helper;
+		$this->template = $template;
+	}
+
+	/**
+	 * Set some template variables for T.I.P.
+	 */
+	public function init_tpl_vars()
+	{
+		$this->template->assign_vars([
+			'S_TOPIC_IMAGE_PREVIEW'		=> $this->helper->is_preview(),
+			'TOPIC_IMAGE_PREVIEW_DIM'	=> $this->config['vse_tip_dim'],
+		]);
 	}
 
 	/**
@@ -61,7 +78,7 @@ class preview implements EventSubscriberInterface
 	 */
 	public function viewforum_row($event)
 	{
-		$this->factory->update_row_data($event);
+		$this->helper->update_row_data($event);
 	}
 
 	/**
@@ -71,7 +88,7 @@ class preview implements EventSubscriberInterface
 	 */
 	public function viewforum_tpl($event)
 	{
-		$this->factory->update_tpl_data($event);
+		$this->helper->update_tpl_data($event);
 	}
 
 	/**
@@ -83,7 +100,7 @@ class preview implements EventSubscriberInterface
 	{
 		if ($event['show_results'] === 'topics' && $this->config->offsetGet('vse_tip_srt'))
 		{
-			$this->factory->update_row_data($event);
+			$this->helper->update_row_data($event);
 		}
 	}
 
@@ -96,7 +113,7 @@ class preview implements EventSubscriberInterface
 	{
 		if ($event['show_results'] === 'topics' && $this->config->offsetGet('vse_tip_srt'))
 		{
-			$this->factory->update_tpl_data($event);
+			$this->helper->update_tpl_data($event);
 		}
 	}
 
@@ -109,7 +126,7 @@ class preview implements EventSubscriberInterface
 	{
 		if ($this->config->offsetGet('vse_tip_pst'))
 		{
-			$this->factory->update_row_data($event);
+			$this->helper->update_row_data($event);
 		}
 	}
 
@@ -122,7 +139,7 @@ class preview implements EventSubscriberInterface
 	{
 		if ($this->config->offsetGet('vse_tip_pst'))
 		{
-			$this->factory->update_tpl_data($event);
+			$this->helper->update_tpl_data($event);
 		}
 	}
 }