+++ tests/event/settings_test.php
@@ -25,12 +25,12 @@ class settings_test extends base
 
 	public function test_construct()
 	{
-		$this->assertInstanceOf('\Symfony\Component\EventDispatcher\EventSubscriberInterface', $this->getEventListener());
+		self::assertInstanceOf('\Symfony\Component\EventDispatcher\EventSubscriberInterface', $this->getEventListener());
 	}
 
 	public function test_getSubscribedEvents()
 	{
-		$this->assertEquals([
+		self::assertEquals([
 			'core.permissions',
 			'core.acp_board_config_edit_add',
 			'core.ucp_prefs_view_data',
@@ -81,23 +81,18 @@ class settings_test extends base
 		$event_data_after = $event->get_data_filtered($event_data);
 		foreach ($event_data as $expected)
 		{
-			$this->assertArrayHasKey($expected, $event_data_after);
+			self::assertArrayHasKey($expected, $event_data_after);
 		}
 		extract($event_data_after, EXTR_OVERWRITE);
 
 		$keys = array_keys($display_vars['vars']);
 
-		$this->assertEquals($expected_keys, $keys);
+		self::assertEquals($expected_keys, $keys);
 	}
 
 	public function update_ucp_data_data()
 	{
 		return [
-			[
-				[],
-				[],
-				['user_vse_tip' => 0],
-			],
 			[
 				['user_vse_tip' => 1],
 				[],
@@ -140,7 +135,7 @@ class settings_test extends base
 
 		$event_data_after = $event->get_data_filtered($event_data);
 
-		$this->assertEquals($expected, $event_data_after['sql_ary']);
+		self::assertEquals($expected, $event_data_after['sql_ary']);
 	}
 
 	public function handle_ucp_data_data()
@@ -253,13 +248,13 @@ class settings_test extends base
 		$listener = $this->getEventListener();
 
 		$this->user->data['user_vse_tip'] = 0;
-		$this->request->expects($this->once())
+		$this->request->expects(self::once())
 			->method('variable')
 			->willReturn($user_vse_tip);
 
 		if (!$submit)
 		{
-			$this->template->expects($this->once())
+			$this->template->expects(self::once())
 				->method('assign_vars')
 				->with([
 					'S_VSE_TIP_NUM'  => '3',
@@ -273,7 +268,7 @@ class settings_test extends base
 
 		$result = $event->get_data_filtered($event_data);
 
-		$this->assertEquals($expected, $result['data']);
+		self::assertEquals($expected, $result['data']);
 	}
 
 	public function add_permissions_test_data()
@@ -310,6 +305,6 @@ class settings_test extends base
 
 		$listener->add_permission($event);
 
-		$this->assertSame($event['permissions'], $expected);
+		self::assertSame($event['permissions'], $expected);
 	}
 }